+++ Celestial.Almanac/Arduino.Cpp/CelestialAlmanac/AstroLib/Saturn.h
@@ -4,9 +4,9 @@
 class Saturn {
   public: // TODO Make those static
     Saturn();
-    float lSaturn(float x);
-    float bSaturn(float x);
-    float rSaturn(float x);
+    double lSaturn(double x);
+    double bSaturn(double x);
+    double rSaturn(double x);
 };    // Warning!!!! Do not forget the semicolon!!
 
 #endif
